+++ HISTORY.md
@@ -67,6 +67,7 @@
 * SciSnap2 extension update, thanks, Eckart!
 * gui: added hidden experimental 'Use CPU for graphics' preference setting for Canvas2D "willReadFrequently" property
 * gui: added configuration object to IDE initialization
+* gui: added configuration options for design, mode, blocksZoom, hideControls and noCloud
 
 ### 2022-11-22
 * morphic: added retina support for inlined worlds, enabling high-res embedding of Snap! components into other websites